% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcPlot.R
\name{fcPlot}
\alias{fcPlot}
\title{Plotly Fold Change plot}
\usage{
fcPlot(
  glmmResult,
  x1Label,
  x2Label,
  pCutoff = 0.01,
  labels,
  useAdjusted = FALSE,
  plotCutoff = 1,
  graphics = "ggplot",
  colours = c("grey", "lightseagreen", "red", "blue"),
  verbose = FALSE
)
}
\arguments{
\item{glmmResult}{output from glmmQvals or glmmSeq}

\item{x1Label}{The name of the first (inner) x parameter}

\item{x2Label}{The name of the second (outer) x parameter}

\item{pCutoff}{The significance cut-off for colour-coding
(default=0.01)}

\item{labels}{Genes to label on plot}

\item{plotCutoff}{Which probes to include by significance cut-off
(default=1 for all markers)}

\item{graphics}{Graphics system to use: "ggplot" or "plotly"}

\item{colours}{Colour vector for significance}

\item{verbose}{Whether to print statistics}

\item{useAdjust}{whether to use adjusted pvalues
(must have q_ columns in glmmResult)}
}
\value{
Returns dataframe with results for gene-wise glm
}
\description{
Plotly Fold Change plot
}
\keyword{hplot}
