% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmSeq.R
\name{glmApply}
\alias{glmApply}
\title{Fit a glm model for an individual gene (no random effect)}
\usage{
glmApply(
  geneList,
  fullFormula,
  data,
  control,
  modelData,
  designMatrix,
  offset,
  ...
)
}
\arguments{
\item{geneList}{List with gene expression and optional dispersion}

\item{fullFormula}{the model formula. For more information of formula
structure see \code{\link[stats:glm]{stats::glm}}}

\item{data}{The sample data or metadata.}

\item{control}{the glm control (default = list()). For more information see
\code{\link[stats:glm]{stats::glm()}}.}

\item{modelData}{Expanded design matrix}

\item{designMatrix}{The design matrix}

\item{offset}{this can be used to specify an a priori known component to be
included in the linear predictor during fitting. For more information see
\code{\link[stats:glm]{stats::glm()}}.}

\item{...}{Other parameters to pass to
\code{\link[stats:glm]{stats::glm()}}}
}
\value{
Returns a GlmmSeq object with results for gene-wise general linear
mixed models
}
\description{
Fit a glm model for an individual gene (no random effect)
}
\keyword{internal}
