% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmGene.R
\name{glmmGene}
\alias{glmmGene}
\title{Glmm for sequencing results of a single gene}
\usage{
glmmGene(
  modelFormula,
  countdata,
  gene,
  metadata,
  id,
  dispersion,
  sizeFactors = NULL,
  reducedFormula = "",
  modelData = NULL,
  control = glmerControl(optimizer = "bobyqa"),
  zeroCount = 0.125,
  removeSingles = TRUE
)
}
\arguments{
\item{modelFormula}{the formula}

\item{countdata}{the sequencing data}

\item{gene}{the gene name}

\item{metadata}{a data frame of sample information}

\item{id}{Column name in metadata which contains the sample IDs to be used
in pairing}

\item{dispersion}{a numeric for the gene dispersion}

\item{sizeFactors}{the gene size factor (default=NULL)}

\item{reducedFormula}{Reduced design formula (default="")}

\item{modelData}{something something}

\item{control}{the glmer control (default=glmerControl(optimizer="bobyqa"))}

\item{zeroCount}{numerical value to offset zeroes for the purpose of log
(default=0.125)}

\item{removeSingles}{whether to remove unpaired individuals (default=TRUE)}
}
\value{
Returns the fit for the general linear mixed model of a single gene
}
\description{
Glmm for sequencing results of a single gene
}
