% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmSeq.R
\docType{class}
\name{GlmmSeq-class}
\alias{GlmmSeq-class}
\title{An S4 class to define the glmmSeq output}
\description{
An S4 class to define the glmmSeq output
}
\section{Slots}{

\describe{
\item{\code{formula}}{The model formula}

\item{\code{stats}}{the stats}

\item{\code{predict}}{The predicted values}

\item{\code{reducedFormula}}{The reduced formula with removed random effects}

\item{\code{modelData}}{the model data}

\item{\code{optinfo}}{the optional info}

\item{\code{errors}}{errors}
}}

