% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmSeq.R
\name{glmmSeq}
\alias{glmmSeq}
\title{Glmm for sequencing results}
\usage{
glmmSeq(
  modelFormula,
  countdata,
  metadata,
  id,
  dispersions,
  sizeFactors = NULL,
  reducedFormula = "",
  modelData = NULL,
  control = glmerControl(optimizer = "bobyqa"),
  cores = detectCores()/2,
  removeSingles = T,
  verbose = T
)
}
\arguments{
\item{modelFormula}{the formula}

\item{countdata}{the sequencing data}

\item{metadata}{a data frame of sample information}

\item{id}{Column name in metadata which contains the sample IDs to be used
in pairing}

\item{dispersions}{a numeric vector of gene dispersions}

\item{sizeFactors}{size factors (default=NULL)}

\item{reducedFormula}{Reduced design fomula (default="")}

\item{modelData}{something something}

\item{control}{the glmer control (default=glmerControl(optimizer="bobyqa"))}

\item{cores}{number of cores to use. Default=detectCores()/2}

\item{removeSingles}{whether to remove unpaired individuals (default=TRUE)}

\item{verbose}{Logical whether to display messaging (default=TRUE)}
}
\value{
Returns dataframe with results for gene-wise glm
}
\description{
Glmm for sequencing results
}
\keyword{hplot}
