% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmSeq_orig.R
\name{glmmSeq_v1}
\alias{glmmSeq_v1}
\title{Glmm for sequencing results}
\usage{
glmmSeq_v1(
  modelFormula,
  countdata,
  metadata,
  id,
  dispersion,
  sizeFactors = NULL,
  reducedFormula = "",
  modelData = NULL,
  designMatrix = NULL,
  control = glmerControl(optimizer = "bobyqa"),
  cores = 1,
  removeDuplicatedMeasures = FALSE,
  removeSingles = FALSE,
  zeroCount = 0.125,
  verbose = TRUE,
  returnList = FALSE,
  progress = TRUE,
  ...
)
}
\arguments{
\item{modelFormula}{the model formula. This must be of the form \code{"~ ..."}
where the structure is assumed to be \code{"counts ~ ..."}. The formula must
include a random effects term. For more information on formula structure
for random effects see \code{\link[lme4:glmer]{lme4::glmer()}}}

\item{countdata}{the sequencing count data matrix with genes in rows and
samples in columns}

\item{metadata}{a dataframe of sample information with variables in columns
and samples in rows}

\item{id}{Used to specify the column in metadata which contains the
sample IDs to be used in repeated samples for random effects. If not
specified, the function defaults to using the variable after the "|" in the
random effects term in the formula.}

\item{dispersion}{a numeric vector of gene dispersion}

\item{sizeFactors}{size factors (default = NULL). If provided the \code{glmer}
offset is set to log(sizeFactors). For more information see``
\code{\link[lme4:glmer]{lme4::glmer()}}}

\item{reducedFormula}{Reduced design formula (default = "")}

\item{modelData}{Expanded design matrix}

\item{designMatrix}{custom design matrix}

\item{control}{the \code{glmer} optimizer control (default =
\code{glmerControl(optimizer = "bobyqa")}). See
\code{\link[lme4:glmerControl]{lme4::glmerControl()}}.}

\item{cores}{number of cores to use. Default = 1.}

\item{removeDuplicatedMeasures}{whether to remove duplicated
conditions/repeated measurements for a given time point (default = FALSE).}

\item{removeSingles}{whether to remove individuals without repeated measures
(default = FALSE)}

\item{zeroCount}{numerical value to offset zeroes for the purpose of log
(default = 0.125)}

\item{verbose}{Logical whether to display messaging (default = TRUE)}

\item{returnList}{Logical whether to return results as a list or \code{glmmSeq}
object (default = FALSE). Useful for debugging.}

\item{progress}{Logical whether to display a progress bar}

\item{...}{Other parameters to pass to
\code{\link[lme4:glmer]{lme4::glmer()}}}
}
\value{
Returns a \code{GlmmSeq} object with results for gene-wise general linear
mixed models or a list of results if \code{returnList} is \code{TRUE}.
}
\description{
Legacy version retained for testing purposes only.
}
\examples{
data(PEAC_minimal_load)
disp <- apply(tpm, 1, function(x) {
(var(x, na.rm = TRUE)-mean(x, na.rm = TRUE))/(mean(x, na.rm = TRUE)**2)
})
MS4A1glmm <- glmmSeq_v1(~ Timepoint * EULAR_6m + (1 | PATID),
                     countdata = tpm["MS4A1", ],
                     metadata = metadata,
                     dispersion = disp["MS4A1"],
                     verbose = FALSE)
names(attributes(MS4A1glmm))
}
