% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmSeq.R
\name{lmerApply}
\alias{lmerApply}
\title{Fit an lmer model for an individual gene}
\usage{
lmerApply(
  geneList,
  fullFormula,
  data,
  control,
  modelData,
  designMatrix,
  offset,
  ...
)
}
\arguments{
\item{geneList}{List with gene expression and dispersion}

\item{fullFormula}{the model formula. For more information of formula
structure see \code{\link[lme4:glmer]{lme4::glmer()}}}

\item{data}{The sample metadata.}

\item{control}{the lmer control (default = lmerControl()). For more information see
\code{\link[lme4:lmerControl]{lme4::lmerControl()}}.}

\item{modelData}{Expanded design matrix}

\item{designMatrix}{The design matrix}

\item{offset}{this can be used to specify an a priori known component to be
included in the linear predictor during fitting. For more information see
\code{\link[lme4:glmer]{lme4::lmer()}}.}

\item{...}{Other parameters to pass to
\code{\link[lme4:glmer]{lme4::lmer()}}}
}
\value{
Returns a list of results for gene-wise linear mixed models
}
\description{
Fit an lmer model for an individual gene
}
\keyword{internal}
