% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maPlot.R
\name{maPlot}
\alias{maPlot}
\title{MA plots}
\usage{
maPlot(
  glmmResult,
  x1Label,
  x2Label,
  pCutoff = 0.01,
  plotCutoff = 1,
  zeroCountCutoff = 50,
  colours = c("grey", "midnightblue", "mediumvioletred", "goldenrod"),
  labels = c(),
  transpose = FALSE,
  useAdjusted = FALSE,
  graphics = "ggplot",
  verbose = FALSE
)
}
\arguments{
\item{glmmResult}{output from glmmQvals or glmmSeq}

\item{x1Label}{The name of the first (inner) x parameter}

\item{x2Label}{The name of the second (outer) x parameter}

\item{pCutoff}{The significance cut-off for colour-coding (default=0.01)}

\item{plotCutoff}{Which probes to include by significance cut-off
(default=1 for all markers)}

\item{colours}{Colour scheme}

\item{labels}{Genes to label on plot}

\item{transpose}{Logical whether FC is time or group based (in this case
x2Label should be the time parameter)}

\item{graphics}{Either "ggplot" or "plotly"}

\item{verbose}{Whether to print statistics}

\item{useAdjust}{whether to use adjusted pvalues
(must have q_ columns in glmmResult)}
}
\value{
List of plot options
}
\description{
MA plots
}
\keyword{hplot}
