% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAC_minimal_load.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Minimal metadata from PEAC}
\format{
A data frame
\describe{
\item{PATID}{Id for matching patients}
\item{Timepoint}{timepoints}
\item{EULAR_6m}{response data}
}
}
\usage{
metadata
}
\description{
Minimal metadata for paired longitudinal response analysis.
}
\keyword{datasets}
