% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldGlmm.R
\name{oldGlmm}
\alias{oldGlmm}
\title{Glmm gene-wise}
\usage{
oldGlmm(
  countdata,
  id,
  time,
  group,
  col,
  dispersions,
  cores = detectCores()/2,
  removeSingles = TRUE
)
}
\arguments{
\item{countdata}{The count data}

\item{id}{A character vector of subjects ids for pairing over time.}

\item{time}{A character vector for the timepoint}

\item{group}{A character vector for the comparison grouping}

\item{col}{The group name to be used}

\item{dispersions}{A numeric vector of gene dispersions}

\item{cores}{The number of cores to use (default=detectCores()/2)}

\item{removeSingles}{Logical whether to use only paired samples
(default=TRUE)}
}
\value{
Returns dataframe with results for gene-wise glm
}
\description{
Paired plots to show differences between groups and over time
}
\keyword{hplot}
