% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedPlot.R
\name{pairedPlot}
\alias{pairedPlot}
\title{Repeated measures plot}
\usage{
pairedPlot(
  object,
  geneName = NULL,
  x1var = NULL,
  x2var = NULL,
  xTitle = NULL,
  yTitle = geneName,
  title = geneName,
  logTransform = is(object, "GlmmSeq"),
  shapes = 21,
  colours = "grey60",
  lineColours = "grey60",
  markerSize = 0.5,
  fontSize = NULL,
  alpha = 0.7,
  addModel = TRUE,
  addPoints = TRUE,
  x2shift = NULL,
  modelSize = 2,
  modelColours = "royalblue",
  modelLineSize = 1,
  modelLineColours = modelColours,
  errorBarLwd = 2.5,
  errorBarLength = 0.05,
  ...
)
}
\arguments{
\item{object}{A glmmSeq object created by
\code{\link[glmmSeq:glmmSeq]{glmmSeq::glmmSeq()}}}

\item{geneName}{The gene/row name to be plotted}

\item{x1var}{The name of the first (inner) x parameter. This must be able
to be paired using the ID.}

\item{x2var}{The name of the second (outer) x parameter}

\item{xTitle}{Title for the x axis}

\item{yTitle}{Title for the y axis}

\item{title}{Plot title. If NULL gene name is used}

\item{logTransform}{Whether to perform a log10 transform on the y axis}

\item{shapes}{The marker shapes (default=19)}

\item{colours}{The marker colours (default='red') as vector or named vector}

\item{lineColours}{The line colours (default='grey60') as vector or named
vector}

\item{markerSize}{Size of markers (default=2)}

\item{fontSize}{Plot font size}

\item{alpha}{Line and marker opacity (default=0.7)}

\item{addModel}{Whether to add the fit model with markers (default=TRUE)}

\item{modelSize}{Size of model points (default=3)}

\item{modelColours}{Colour of model fit markers (default="black") as vector
or named vector}

\item{modelLineSize}{Size of model points (default=1) as vector or named
vector}

\item{modelLineColours}{Colour of model fit lines.}

\item{errorBarLwd}{Line width of error bars}

\item{errorBarLength}{Head width of error bars}

\item{...}{Other parameters to pass to
\code{\link[graphics:plot]{graphics::plot()}}}

\item{IDColumn}{Column name of sample IDs for pairing}

\item{pairedOnly}{Logical whether to only plot paired samples (default=TRUE)}
}
\value{
Returns a paired plot for matched samples
}
\description{
Plot to show differences between groups over time using base graphics.
}
\examples{
data(PEAC_minimal_load)

disp <- apply(tpm, 1, function(x){
  (var(x, na.rm=TRUE)-mean(x, na.rm=TRUE))/(mean(x, na.rm=TRUE)**2)
})

MS4A1glmm <- glmmSeq2(~ Timepoint * EULAR_6m + (1 | PATID),
                     countdata = tpm['MS4A1', ],
                     metadata = metadata,
                     dispersion = disp['MS4A1'])

pairedPlot(object=MS4A1glmm,
           geneName = 'MS4A1',
           x1var = 'Timepoint',
           x2var='EULAR_6m')
}
