% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAC_minimal_load.R
\docType{data}
\name{tpm}
\alias{tpm}
\title{TPM count data from PEAC}
\format{
An object of class \code{data.frame} with 3000 rows and 162 columns.
}
\usage{
tpm
}
\description{
A dataset containing the count data
}
\keyword{datasets}
