% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcPlot.R
\name{fcPlot}
\alias{fcPlot}
\title{Plotly Fold Change plot}
\usage{
fcPlot(
  glmmResults,
  groupCol,
  colourCutoff = 0.01,
  labels,
  useAdjusted = F,
  plotCutoff = 0.05
)
}
\arguments{
\item{glmmResults}{output from glmmQvals or glmmSeq}

\item{groupCol}{The grouping column}

\item{colourCutoff}{The significance cutoff for colour-coding (default=0.01)}

\item{labels}{Genes to label on plot}

\item{plotCutoff}{Which probes to include by significance cutoff
(default=0.05)}

\item{useAdjust}{whether to use adjusted pvalues
(must have q_ columns in glmmResults)}
}
\value{
Returns dataframe with results for gene-wise glm
}
\description{
Plotly Fold Change plot
}
\keyword{hplot}
