% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmQvals.R
\name{glmmQvals}
\alias{glmmQvals}
\title{Glmm Sequencing qvalues}
\usage{
glmmQvals(result, cutoff = 0.05, pi0 = NULL, verbose = TRUE)
}
\arguments{
\item{result}{output from glmmSeq}

\item{cutoff}{Prints a table showing the number of probes considered
significant by the pvalue cut-off (default=0.05)}

\item{pi0}{It is recommended not to input an estimate of pi0. Experienced
users can use their own methodology to estimate the proportion of true nulls
or set it equal to 1 for the BH procedure (default = NULL).}

\item{verbose}{Logical whether to print the number of significant probes
(default=TRUE)}
}
\value{
Returns dataframe with results for gene-wise glm with qvalue columns
}
\description{
Add qvalue columns to the glmmSeq dataframe
}
