% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmSeq.R
\docType{class}
\name{GlmmSeq-class}
\alias{GlmmSeq-class}
\title{An S4 class to define the glmmSeq output}
\description{
An S4 class to define the glmmSeq output
}
\section{Slots}{

\describe{
\item{\code{formula}}{The model formula}

\item{\code{stats}}{the statistics from the glmm fit}

\item{\code{predict}}{The predicted interception values}

\item{\code{reducedFormula}}{The reduced formula with removed random effects}

\item{\code{countdata}}{The input expression data}

\item{\code{metadata}}{The input metadata}

\item{\code{modelData}}{the model data for the glmm}

\item{\code{optInfo}}{Information on whether the model was singular or converged}

\item{\code{errors}}{Any errors}

\item{\code{variables}}{The variables used in the formula}
}}

