% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmSeq.R
\docType{class}
\name{GlmmSeq-class}
\alias{GlmmSeq-class}
\title{An S4 class to define the glmmSeq output}
\description{
An S4 class to define the glmmSeq output
}
\section{Slots}{

\describe{
\item{\code{call}}{The matched call}

\item{\code{formula}}{The model formula}

\item{\code{stats}}{the statistics from the glmm fit}

\item{\code{predict}}{The predicted interception values}

\item{\code{reducedFormula}}{The reduced formula with removed random effects}

\item{\code{countdata}}{The input expression data}

\item{\code{metadata}}{The input metadata}

\item{\code{dispersion}}{Input dispersion vector}

\item{\code{modelData}}{Model data for predictions}

\item{\code{offset}}{Model offset}

\item{\code{designMatrix}}{Model design matrix}

\item{\code{optInfo}}{Information on whether the model was singular or converged}

\item{\code{errors}}{Any errors}

\item{\code{vars}}{List of variables stored from the original call}
}}

