% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmSeq.R
\docType{class}
\name{lmmSeq-class}
\alias{lmmSeq-class}
\title{An S4 class to define the lmmSeq output}
\description{
An S4 class to define the lmmSeq output
}
\section{Slots}{

\describe{
\item{\code{info}}{List including matched call, offset, designMatrix}

\item{\code{formula}}{The model formula}

\item{\code{stats}}{Statistics from fitted models}

\item{\code{predict}}{Predicted values}

\item{\code{reducedFormula}}{The reduced formula with removed random effects}

\item{\code{maindata}}{The input expression data with variables in rows}

\item{\code{metadata}}{The input metadata}

\item{\code{modelData}}{Model data for predictions}

\item{\code{optInfo}}{Information on whether the model was singular or converged}

\item{\code{errors}}{Any errors}

\item{\code{vars}}{List of variables stored from the original call}
}}

