% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmSeq2.R
\name{lmmSeq2}
\alias{lmmSeq2}
\title{Linear mixed models for data matrix}
\usage{
lmmSeq2(
  modelFormula,
  maindata,
  metadata,
  id = NULL,
  offset = NULL,
  reducedFormula = "",
  modelData = NULL,
  designMatrix = NULL,
  control = lmerControl(),
  cores = 1,
  removeSingles = FALSE,
  verbose = TRUE,
  returnList = FALSE,
  progress = FALSE,
  ...
)
}
\arguments{
\item{modelFormula}{the model formula. This must be of the form \code{"~ ..."}
where the structure is assumed to be \code{"gene ~ ..."}. The formula must
include a random effects term. See formula structure for random effects in
\code{\link[lme4:lmer]{lme4::lmer()}}}

\item{maindata}{data matrix with genes in rows and samples in columns}

\item{metadata}{a dataframe of sample information with variables in columns
and samples in rows}

\item{id}{Optional. Used to specify the column in metadata which contains the
sample IDs to be used in repeated samples for random effects. If not
specified, the function defaults to using the variable after the "|" in the
random effects term in the formula.}

\item{offset}{Vector containing model offsets (default = NULL). If provided
the \code{lmer()} offset is set to \code{offset}. See
\code{\link[lme4:lmer]{lme4::lmer()}}}

\item{reducedFormula}{Optional design formula without random effects. If not
given, it is automatically generated by removing the random effects from
the main formula. Used to calculate confidence intervals for final fitted
models on each gene for plotting purposes.}

\item{modelData}{Optional dataframe. Default is generated by call to
\code{expand.grid} using levels of variables in the formula. Used to calculate
model predictions for plotting.}

\item{designMatrix}{Optional custom design matrix generated by call to
\code{model.matrix} using \code{modelData} and \code{reducedFormula}. Used to calculate
model predictions for plotting.}

\item{control}{the \code{lmer} optimizer control (default = \code{lmerControl()}). See
\code{\link[lme4:lmerControl]{lme4::lmerControl()}}.}

\item{cores}{number of cores to use for parallelisation. Default = 1.}

\item{removeSingles}{whether to remove individuals with no repeated measures
(default = FALSE)}

\item{verbose}{Logical whether to display messaging (default = TRUE)}

\item{returnList}{Logical whether to return results as a list or lmmSeq
object (default = FALSE). Helpful for debugging.}

\item{progress}{Logical whether to display a progress bar}

\item{...}{Other parameters passed to \code{\link[lme4:lmer]{lme4::lmer()}}}
}
\value{
Returns an S4 class \code{lmmSeq} object with results for gene-wise
linear mixed models or a list of results if \code{returnList} is \code{TRUE}.
}
\description{
Experimental version to speed up Wald type 2 Chi-square test by vectorising
code from car::Anova
}
\examples{
data(PEAC_minimal_load)
logtpm <- log2(tpm +1)
lmmtest <- lmmSeq(~ Timepoint * EULAR_6m + (1 | PATID),
                     maindata = logtpm["MS4A1", ],
                     metadata = metadata,
                     verbose = FALSE)
names(attributes(lmmtest))
}
