% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maPlot.R
\name{maPlotly}
\alias{maPlotly}
\title{MA plots}
\usage{
maPlotly(
  glmmResults,
  groupCol,
  colourCutoff = 0.01,
  plotCutoff = 0.05,
  colours = c("grey", "orange", "red", "blue"),
  labels = c(),
  transpose = FALSE,
  useAdjusted = FALSE
)
}
\arguments{
\item{glmmResults}{output from glmmQvals or glmmSeq}

\item{groupCol}{The grouping column}

\item{colourCutoff}{The significance cut-off for colour-coding (default=0.01)}

\item{plotCutoff}{Which probes to include by significance cut-off
(default=0.05)}

\item{colours}{Colour scheme}

\item{labels}{Genes to label on plot}

\item{transpose}{Logical whether FC is time or group based (in this case
groupCol should be the time parameter)}

\item{useAdjust}{whether to use adjusted pvalues
(must have q_ columns in glmmResults)}
}
\value{
List of plot options
}
\description{
MA plots
}
\keyword{hplot}
