% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlot.R
\name{modelPlot}
\alias{modelPlot}
\title{Model plot}
\usage{
modelPlot(result, gene, pch = 21, bg = c("red", "blue"), graphics = "ggplot")
}
\arguments{
\item{result}{A glmmSeq object}

\item{gene}{Gene name to plot}

\item{pch}{The marker shape (default=21)}

\item{bg}{The marker colour, default=c('red', 'blue')}

\item{graphics}{Which graphic system to use (options = "base" or "ggplot")}
}
\value{
Returns a model plot
}
\description{
Model plots to show the overall differences between groups and over time
}
\keyword{hplot}
