% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlot.R
\name{modelPlot}
\alias{modelPlot}
\title{Model plot}
\usage{
modelPlot(
  glmmResult,
  geneName,
  x1Label = "Timepoint",
  x2Label,
  xTitle = NULL,
  logTransform = FALSE,
  shapes = 21,
  colours = c("red", "blue"),
  graphics = "ggplot"
)
}
\arguments{
\item{glmmResult}{A glmmSeq object}

\item{geneName}{Gene name to plot}

\item{x1Label}{The name of the first (inner) x parameter}

\item{x2Label}{The name of the second (outer) x parameter}

\item{xTitle}{Title for the x axis}

\item{logTransform}{Whether to perform a log10 transform on the y axis}

\item{shapes}{The marker shape (default=21)}

\item{colours}{The marker colour, default=c('red', 'blue')}

\item{graphics}{Which graphic system to use (options = "base" or "ggplot")}
}
\value{
Returns a model plot
}
\description{
Model plots to show the overall differences between groups and over time
}
