% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedPlots.R
\name{pairedPlot}
\alias{pairedPlot}
\title{Paired plots}
\usage{
pairedPlot(
  glmmResult,
  geneName = NULL,
  x1Label = NULL,
  x2Label = NULL,
  IDColumn = "ID",
  xTitle = NULL,
  logTransform = FALSE,
  shapes = 21,
  colours = "red",
  lineColour = "grey60",
  alpha = 0.7,
  pairedOnly = TRUE,
  title = "",
  graphics = "base",
  addModel = TRUE,
  modSize = 1.5,
  modColour = "black",
  modLineColour = "black",
  addBox = FALSE,
  addViolins = TRUE,
  violinWidth = 0.5,
  ...
)
}
\arguments{
\item{glmmResult}{A GlmmSeq object from \code{glmmSeq}.}

\item{geneName}{The gene name to be plotted.}

\item{x1Label}{The name of the first (inner) x parameter. This must be able
to be paired using the ID.}

\item{x2Label}{The name of the second (outer) x parameter}

\item{xTitle}{Title for the x axis}

\item{logTransform}{Whether to perform a log10 transform on the y axis}

\item{shapes}{The marker shapes (default=21)}

\item{colours}{The marker colours (default='red')}

\item{lineColour}{The line colours (default='grey60')}

\item{alpha}{Line and marker opacity (default=0.7)}

\item{pairedOnly}{Logical whether to only plot paired samples (default=TRUE)}

\item{title}{Plot title}

\item{graphics}{Which graphic system to use (options = "base" or "ggplot")}

\item{addModel}{Whether to add the fit model with markers}

\item{modSize}{Size of model points and line}

\item{modColour}{Colour of model fit markers}

\item{modLineColour}{Colour of model fit lines}

\item{addBox}{Logical whether to add boxplots.}

\item{addViolins}{Logical whether to add half violins (ggplot only)}

\item{violinWidth}{Width of violin plots}
}
\value{
Returns a paired plot
}
\description{
Paired plots to show differences between groups and over time
}
