% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedPlots.R
\name{pairedPlot}
\alias{pairedPlot}
\title{Paired plots}
\usage{
pairedPlot(
  id,
  time,
  gene_exp,
  f,
  dispersion,
  datatype,
  plottype,
  plab = "drug",
  pch = 21,
  bg = "red",
  pairedonly = T,
  xlab = "",
  ylab = "",
  title = "",
  graphics = "base",
  addBox = TRUE,
  ...
)
}
\arguments{
\item{id}{A character vector of subjects ids for pairing over time.}

\item{time}{The time point}

\item{gene_exp}{The gene expression}

\item{dispersion}{The dispersion}

\item{datatype}{The type of expression data. Options include c("VST", "TPM")}

\item{plottype}{The plot type. Options "1"=plot all samples, "2"=plot mean}

\item{pch}{The marker shape (default=21)}

\item{pairedonly}{Logical whether to only plot paired samples (default=T)}

\item{xlab}{Character vector for the label on the x axis}

\item{graphics}{Which graphic system to use (options = "base" or "ggplot")}

\item{x}{A character vector for groups}

\item{bg='red'}{The marker colour (default='red')}
}
\value{
Returns a paired plot
}
\description{
Paired plots to show differences between groups and over time
}
\keyword{hplot}
