% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEAC_minimal_load.R
\docType{data}
\name{tpm}
\alias{tpm}
\title{TPM count data from PEAC}
\format{
An object of class \code{data.frame} with 50 rows and 149 columns.
}
\usage{
tpm
}
\description{
Transcripts Per Million (TPM) count data for PEAC synovial
biopsies.
}
\keyword{datasets}
